Deprecate escaping in a backward-compatible way.
