+++ module/VuFind/src/VuFind/View/Helper/Root/RecordLink.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Record link view helper
+ * Record link view helper (DEPRECATED -- use RecordLinker instead)
  *
  * PHP version 7
  *
@@ -28,7 +28,7 @@
 namespace VuFind\View\Helper\Root;
 
 /**
- * Record link view helper
+ * Record link view helper (DEPRECATED -- use RecordLinker instead)
  *
  * @category VuFind
  * @package  View_Helpers
@@ -38,23 +38,6 @@ namespace VuFind\View\Helper\Root;
  */
 class RecordLink extends \Laminas\View\Helper\AbstractHelper
 {
-    /**
-     * Record router
-     *
-     * @var \VuFind\Record\Router
-     */
-    protected $router;
-
-    /**
-     * Constructor
-     *
-     * @param \VuFind\Record\Router $router Record router
-     */
-    public function __construct(\VuFind\Record\Router $router)
-    {
-        $this->router = $router;
-    }
-
     /**
      * Given an array representing a related record (which may be a bib ID or OCLC
      * number), this helper renders a URL linking to that record.
@@ -67,61 +50,20 @@ class RecordLink extends \Laminas\View\Helper\AbstractHelper
      */
     public function related($link, $escape = true, $source = DEFAULT_SEARCH_BACKEND)
     {
-        $urlHelper = $this->getView()->plugin('url');
-        $baseUrl = $urlHelper($this->getSearchActionForSource($source));
-        switch ($link['type']) {
-        case 'bib':
-            $url = $baseUrl
-                . '?lookfor=' . urlencode($link['value'])
-                . '&type=id&jumpto=1';
-            break;
-        case 'dlc':
-            $url = $baseUrl
-                . '?lookfor=' . urlencode('"' . $link['value'] . '"')
-                . '&type=lccn&jumpto=1';
-            break;
-        case 'isn':
-            $url = $baseUrl
-                . '?join=AND&bool0[]=AND&lookfor0[]=%22'
-                . urlencode($link['value'])
-                . '%22&type0[]=isn&bool1[]=NOT&lookfor1[]=%22'
-                . urlencode($link['exclude'])
-                . '%22&type1[]=id&sort=title&view=list';
-            break;
-        case 'oclc':
-            $url = $baseUrl
-                . '?lookfor=' . urlencode($link['value'])
-                . '&type=oclc_num&jumpto=1';
-            break;
-        case 'title':
-            $url = $baseUrl
-                . '?lookfor=' . urlencode($link['value'])
-                . '&type=title';
-            break;
-        default:
-            throw new \Exception('Unexpected link type: ' . $link['type']);
-        }
-
+        // Call parent helper, omitting $escape param:
+        $value = $this->__call(__METHOD__, [$link, $source]);
         $escapeHelper = $this->getView()->plugin('escapeHtml');
-        return $escape ? $escapeHelper($url) : $url;
+        return $escape ? $escapeHelper($value) : $value;
     }
 
     /**
-     * Given a record driver, get a URL for that record.
+     * Magic method to proxy recordLinker functionality.
      *
-     * @param \VuFind\RecordDriver\AbstractBase|string $driver Record driver
-     * representing record to link to, or source|id pipe-delimited string
-     * @param string                                   $action Record
-     * action to access
-     *
-     * @return string
+     * @return mixed
      */
-    public function getActionUrl($driver, $action)
+    public function __call($method, $args)
     {
-        // Build the URL:
-        $urlHelper = $this->getView()->plugin('url');
-        $details = $this->router->getActionRouteDetails($driver, $action);
-        return $urlHelper($details['route'], $details['params'] ?: []);
+        return $this->getView()->plugin('recordLinker')->$method(...$args);
     }
 
     /**
@@ -135,7 +77,7 @@ class RecordLink extends \Laminas\View\Helper\AbstractHelper
      */
     public function getHoldUrl($url, $includeAnchor = true)
     {
-        return $this->getRequestUrl($url, $includeAnchor);
+        return $this->__call('getRequestUrl', func_get_args());
     }
 
     /**
@@ -148,82 +90,10 @@ class RecordLink extends \Laminas\View\Helper\AbstractHelper
      */
     public function getRequestUrl($url, $includeAnchor = true)
     {
-        if (is_array($url)) {
-            // Assemble URL string from array parts:
-            $source = $url['source'] ?? DEFAULT_SEARCH_BACKEND;
-            $finalUrl
-                = $this->getActionUrl("{$source}|" . $url['record'], $url['action']);
-            if (isset($url['query'])) {
-                $finalUrl .= '?' . $url['query'];
-            }
-            if (isset($url['anchor']) && $includeAnchor) {
-                $finalUrl .= $url['anchor'];
-            }
-        } else {
-            // If URL is already a string but we don't want anchors, strip
-            // the anchor now:
-            if (!$includeAnchor) {
-                [$finalUrl] = explode('#', $url);
-            } else {
-                $finalUrl = $url;
-            }
-        }
+        $finalUrl = $this->__call(__METHOD__, func_get_args());
         // Make sure everything is properly HTML encoded:
         $escaper = $this->getView()->plugin('escapehtml');
         return $escaper($finalUrl);
-    }
-
-    /**
-     * Given a record driver, get a URL for that record.
-     *
-     * @param \VuFind\RecordDriver\AbstractBase|string $driver Record driver
-     * representing record to link to, or source|id pipe-delimited string
-     * @param string                                   $tab    Optional record
-     * tab to access
-     * @param array                                    $query  Optional query params
-     *
-     * @return string
-     */
-    public function getTabUrl($driver, $tab = null, $query = [])
-    {
-        // Build the URL:
-        $urlHelper = $this->getView()->plugin('url');
-        $details = $this->router->getTabRouteDetails($driver, $tab, $query);
-        return $urlHelper(
-            $details['route'],
-            $details['params'],
-            $details['options'] ?? []
-        );
-    }
-
-    /**
-     * Get the default URL for a record.
-     *
-     * @param \VuFind\RecordDriver\AbstractBase|string $driver Record driver
-     * representing record to link to, or source|id pipe-delimited string
-     *
-     * @return string
-     */
-    public function getUrl($driver)
-    {
-        // Display default tab by default:
-        return $this->getTabUrl($driver);
-    }
-
-    /**
-     * Given a record driver, generate HTML to link to the record from breadcrumbs.
-     *
-     * @param \VuFind\RecordDriver\AbstractBase $driver Record to link to.
-     *
-     * @return string
-     */
-    public function getBreadcrumb($driver)
-    {
-        $truncateHelper = $this->getView()->plugin('truncate');
-        $escapeHelper = $this->getView()->plugin('escapeHtml');
-        return '<a href="' . $this->getUrl($driver) . '">' .
-            $escapeHelper($truncateHelper($driver->getBreadcrumb(), 30))
-            . '</a>';
     }
 
     /**
@@ -235,12 +105,7 @@ class RecordLink extends \Laminas\View\Helper\AbstractHelper
      */
     public function getChildRecordSearchUrl($driver)
     {
-        $urlHelper = $this->getView()->plugin('url');
-        $route = $this->getSearchActionForSource($driver->getSourceIdentifier());
-        $url = $urlHelper($route)
-            . '?lookfor='
-            . urlencode(addcslashes($driver->getUniqueID(), '"'))
-            . '&type=ParentID';
+        $url = $this->__call(__METHOD__, func_get_args());
         // Make sure everything is properly HTML encoded:
         $escaper = $this->getView()->plugin('escapehtml');
         return $escaper($url);
@@ -255,47 +120,9 @@ class RecordLink extends \Laminas\View\Helper\AbstractHelper
      */
     public function getVersionsSearchUrl($driver)
     {
-        $route = $this->getVersionsActionForSource($driver->getSourceIdentifier());
-        if (false === $route) {
-            return '';
-        }
-
-        $urlParams = [
-            'id' => $driver->getUniqueID(),
-            'keys' => $driver->tryMethod('getWorkKeys', [], [])
-        ];
-
-        $urlHelper = $this->getView()->plugin('url');
-        $url = $urlHelper($route, [], ['query' => $urlParams]);
+        $url = $this->__call(__METHOD__, func_get_args());
         // Make sure everything is properly HTML encoded:
         $escaper = $this->getView()->plugin('escapehtml');
         return $escaper($url);
-    }
-
-    /**
-     * Given a record source ID, return the route name for searching its backend.
-     *
-     * @param string $source Record source identifier.
-     *
-     * @return string
-     */
-    protected function getSearchActionForSource($source)
-    {
-        $optionsHelper = $this->getView()->plugin('searchOptions');
-        return $optionsHelper($source)->getSearchAction();
-    }
-
-    /**
-     * Given a record source ID, return the route name for version search with its
-     * backend.
-     *
-     * @param string $source Record source identifier.
-     *
-     * @return string|bool
-     */
-    protected function getVersionsActionForSource($source)
-    {
-        $optionsHelper = $this->getView()->plugin('searchOptions');
-        return $optionsHelper($source)->getVersionsAction();
     }
 }
