+++ themes/bootstrap3/templates/RecordDriver/DefaultRecord/data-containerTitle.phtml
@@ -2,9 +2,9 @@
   $containerSource = $this->driver->getSourceIdentifier();
   $containerID = $this->driver->getContainerRecordID();
   $ref = $this->driver->getContainerReference();
-  $link = $containerID
-    ? $this->recordLink()->getUrl("$containerSource|$containerID")
+  $linkEsc = $containerID
+    ? $this->escapeHtmlAttr($this->recordLinker()->getUrl("$containerSource|$containerID"))
     : $this->record($this->driver)->getLink('journaltitle', $data);
 ?>
-<a href="<?=$link?>"><?=$this->escapeHtml($data)?></a>
+<a href="<?=$linkEsc?>"><?=$this->escapeHtml($data)?></a>
 <?=empty($ref) ? '' : $this->escapeHtml($ref)?>
