+++ themes/root/theme.config.php
@@ -47,7 +47,8 @@ return [
             'VuFind\View\Helper\Root\Recommend' => 'Laminas\ServiceManager\Factory\InvokableFactory',
             'VuFind\View\Helper\Root\Record' => 'VuFind\View\Helper\Root\RecordFactory',
             'VuFind\View\Helper\Root\RecordDataFormatter' => 'VuFind\View\Helper\Root\RecordDataFormatterFactory',
-            'VuFind\View\Helper\Root\RecordLink' => 'VuFind\View\Helper\Root\RecordLinkFactory',
+            'VuFind\View\Helper\Root\RecordLink' => 'Laminas\ServiceManager\Factory\InvokableFactory',
+            'VuFind\View\Helper\Root\RecordLinker' => 'VuFind\View\Helper\Root\RecordLinkerFactory',
             'VuFind\View\Helper\Root\Relais' => 'VuFind\View\Helper\Root\RelaisFactory',
             'VuFind\View\Helper\Root\Related' => 'VuFind\View\Helper\Root\RelatedFactory',
             'VuFind\View\Helper\Root\RenderArray' => 'Laminas\ServiceManager\Factory\InvokableFactory',
@@ -119,6 +120,7 @@ return [
             'record' => 'VuFind\View\Helper\Root\Record',
             'recordDataFormatter' => 'VuFind\View\Helper\Root\RecordDataFormatter',
             'recordLink' => 'VuFind\View\Helper\Root\RecordLink',
+            'recordLinker' => 'VuFind\View\Helper\Root\RecordLinker',
             'relais' => 'VuFind\View\Helper\Root\Relais',
             'related' => 'VuFind\View\Helper\Root\Related',
             'renderArray' => 'VuFind\View\Helper\Root\RenderArray',
